schema and CLI updates
